Modularize tests more
Updates
