+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-features-test-js-alignments-track-colors-by-tag-updates-object-and-layout-1-snap.png
