+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-features-test-js-alignments-track-opens-the-track-menu-and-enables-soft-clipping-1-snap.png
